Countripedia/JS/Api/Bootstrap
This is a simple application for collecting information about the countries of the world. You can search through the search bar or through the buttons that are placed and show the countries according to which continent they are. This application is made using javascript, bootstrap, html and css. Included is an api for all countries in the world.